ENSWEB-1995: Remove range "catastrophe" code.
A "catastrophe" is when a range slider suddenly decides that the
correct cohromosome to display the range for is not the one it has
up to that point, but some other, due to the overwhelming number of
rows on that chromosome.

This will never get activated in the current code as all ranges are
hardwired. It has been broken for a while due to refactoring elsewhere.

Longer term, we need this back in to handle the general case, but it
would need completely rewriting anyway. So the old code is dumped for now.