Fix typo in a screen command output by the script
The output of the script includes a `screen` command to run, but the command as specified uses `screen -s` (lowercase "S"), which is the option to set the shell in a new session. I think it's probably meant to be `screen -S` (capital "S"), which is the option to name the new session.